refactor: ui 컴포넌트 이름 변경
