+++ app/src/main/java/woowacourse/shopping/ProductListActivity.kt
@@ -7,21 +7,10 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.compose.foundation.border
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.unit.dp
 import woowacourse.shopping.ui.ProductListScreen
+import woowacourse.shopping.ui.component.MoreButton
 import woowacourse.shopping.ui.pagination.ProductDataLoadStateHolder
 import woowacourse.shopping.ui.theme.AndroidShoppingTheme
 
@@ -53,25 +42,5 @@ class ProductListActivity : ComponentActivity() {
                 }
             }
         }
-    }
-}
-
-@Composable
-fun MoreButton(onClick: () -> Unit) {
-    Button(
-        onClick = onClick,
-        colors =
-            ButtonDefaults.buttonColors(
-                contentColor = MaterialTheme.colorScheme.onBackground,
-                containerColor = Color.Transparent,
-            ),
-        modifier =
-            Modifier.fillMaxWidth().padding(vertical = 10.dp, horizontal = 10.dp).border(
-                width = 1.dp,
-                color = Color.Black,
-                shape = RoundedCornerShape(10.dp),
-            ),
-    ) {
-        Text("더보기")
     }
 }
