+++ app/src/main/java/woowacourse/shopping/ShoppingCartActivity.kt
@@ -12,8 +12,8 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
-import woowacourse.shopping.ui.AnglePager
 import woowacourse.shopping.ui.ShoppingCartScreen
+import woowacourse.shopping.ui.component.PageNavigation
 import woowacourse.shopping.ui.pagination.ShoppingCartPaginationStateHolder
 import woowacourse.shopping.ui.theme.AndroidShoppingTheme
 
@@ -41,7 +41,7 @@ class ShoppingCartActivity : ComponentActivity() {
                         shoppingCartItems = shoppingCartRepository.getShoppingItems()
                     },
                 ) {
-                    AnglePager(
+                    PageNavigation(
                         currentPage = shoppingCartPaginationStateHolder.currentPage,
                         canMoveToPreviousPage = shoppingCartPaginationStateHolder.canMoveToPreviousPage(),
                         canMoveToNextPage = shoppingCartPaginationStateHolder.canMoveToNextPage(),
