+++ app/src/main/java/woowacourse/shopping/ui/component/MoreButton.kt
@@ -1,0 +1,34 @@
+package woowacourse.shopping.ui.component
+
+import androidx.compose.foundation.border
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun MoreButton(onClick: () -> Unit) {
+    Button(
+        onClick = onClick,
+        colors =
+            ButtonDefaults.buttonColors(
+                contentColor = MaterialTheme.colorScheme.onBackground,
+                containerColor = Color.Transparent,
+            ),
+        modifier =
+            Modifier.fillMaxWidth().padding(vertical = 10.dp, horizontal = 10.dp).border(
+                width = 1.dp,
+                color = Color.Black,
+                shape = RoundedCornerShape(10.dp),
+            ),
+    ) {
+        Text("더보기")
+    }
+}
